MUVT added
This tool (MUlti-objective eValuation Tool; MUVT) enables a multi-objective analysis that allows for a quick and intuitive evaluation of the impact that a specific agro-input (e.g., seasonal irrigation volume, nitrogen fertilization, increase in atmospheric CO2, etc.) has on various variables of the cropping system (e.g., yield, income, management costs, efficiency in the use of the specific input, and so on). In this way, a user can assess the effect of their crop management decisions on variables such as productive, environmental, and economic performance and act accordingly. Furthermore, to support these decisions, the tool provides a numerical assessment based on the behavior of the entire cropping system (integrating the response of all the analyzed variables) in relation to different amounts of agro-inputs, also modulating the weight that each analyzed variable can have in the decision-making process.